Print tuples when compiled with DBT_TRACE or DBT_TRACE_ALL
Compiling with flag DBT_TRACE was broken because tuples were not being parsed
before being printed. This modification adds a method called print_tuple that
parses the current tuple and prints it to stdout. Tuples are only printed to the
standard output when compiled with DBT_TRACE or DBT_TRACE_ALL.

Signed-off-by: Edson Lucas <edson.lucas@uni-passau.de>